Speed-up help option of CLI with lazy inlined imports
