+++ pyproject.toml
@@ -110,6 +110,9 @@ ignore = [
     "SIM103",   # Don't recommend returning the condition directly
 ]
 
+[tool.ruff.lint.per-file-ignores]
+"_cli.py" = ["PLC0415"]  # Ignore imports that are not at the top-level
+
 
 [tool.pytest.ini_options]
 minversion = "8"
